Начало рефакторинга database.js
